requirements.txt
torch>=2.4.0
torchvision>=0.19.0
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0

